pyproject.toml
[project]
name = "context-confusion"
version = "0.1.0"
description = "Demo showcasing context confusion in tool-use agents and how to solve it with LangSmith evaluations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Christina Astoria", email = "christina@example.com"}
]

dependencies = [
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.26.0",
    "openevals>=0.1.0",
    "python-dotenv>=1.0.0",
    "deepagents>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["context_confusion*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

